pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "auto-register"
version = "2.0.0"
description = "高度可配置的网站自动注册框架，支持可插拔的邮箱提供者、站点策略和输出格式。"
requires-python = ">=3.10"

dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "playwright>=1.40",
]

[project.optional-dependencies]
stealth = ["playwright-stealth>=1.0"]

[project.scripts]
auto-register = "cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]
